Rag_chatbot/rag_system/requirements.txt
# Core RAG dependencies
langchain==0.1.0
langchain-community==0.0.13
langchain-anthropic==0.1.1

# Vector stores
chromadb==0.4.22
faiss-cpu==1.7.4

# Embeddings
anthropic==0.18.1
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
tiktoken==0.5.2

# Optional: Graph DB
neo4j==5.15.0

# Development
jupyter==1.0.0
ipython==8.19.0
